Move dynamics code around a bit
